# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateTenantRequest:
    properties:
      companyName:
        example: Acme Corp
        type: string
      domain:
        example: acme.com
        type: string
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        type: string
    required:
    - companyName
    - domain
    - email
    - name
    type: object
...
  dto.TenantDTO:
    properties:
      active:
        example: true
        type: boolean
      companyName:
        example: Acme Corp
        type: string
      createdAt:
        example: "2023-01-01T00:00:00Z"
        type: string
      domain:
        example: acme.com
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: John Doe
        type: string
      updatedAt:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  dto.UpdateTenantRequest:
    properties:
      active:
        example: false
        type: boolean
      companyName:
        example: Acme Inc
        type: string
      domain:
        example: acme-inc.com
        type: string
      email:
        example: john.smith@example.com
        type: string
      name:
        example: John Smith
        type: string
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: Human Resource Information System API
  title: HRIS API
  version: "1.0"
paths:
...
  /tenants:
    get:
      consumes:
      - application/json
      description: Get all tenants
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.TenantDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tenants
      tags:
      - tenants
    post:
      consumes:
      - application/json
      description: Create a new tenant (Admin only)
      parameters:
      - description: Tenant data
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.TenantDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new tenant
      tags:
      - tenants
  /tenants/{id}:
    get:
      consumes:
      - application/json
      description: Get tenant by ID
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TenantDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get tenant by ID
      tags:
      - tenants
    put:
      consumes:
      - application/json
      description: Update a tenant (Admin only)
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant data
        in: body
        name: tenant
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTenantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TenantDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a tenant
      tags:
      - tenants
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"